% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualdiagmvls.R
\name{visualdiagmvls}
\alias{visualdiagmvls}
\title{mvls}
\usage{
visualdiagmvls(mvls)
}
\arguments{
\item{mvls}{A object class mvls}
}
\value{
Plot
}
\description{
Takes a mvls object and give a plot about cluster distribution
}
\author{
Lorenzo Querci <lorenzo.querci@studio.unibo.it>
}
